+++ setup/workspace_setup.py
@@ -1,7 +1,7 @@
 import bpy
 from pathlib import Path
 from typing import List, Sequence, Tuple
-from .. constants import APP_TEMPLATES_DIR
+from .. constants import APP_TEMPLATES_DIR, STORYBOARD_TEMPLATE_BLEND, DUAL_STORYBOARD_TEMPLATE_BLEND
 from bpy.types import (
     Window,
     WindowManager,
@@ -10,10 +10,71 @@ from bpy.types import (
 from .. import fn
 
 def set_sidebar(area=None):
+    prefs = fn.get_addon_prefs()
+
+    if not area:
+        area = bpy.context.area
+
+    context = bpy.context
+    window = bpy.context.window
+
+    if area == None:
+        ## Return biggest 3D area
+        areas_3d = [(w, a) for w in bpy.context.window_manager.windows for a in w.screen.areas if a.type == 'VIEW_3D']
+        if not areas_3d:
+            print('Could not found area ! skip "set sidebar" operation')
+            return
+        areas_3d.sort(key=lambda x: x[1].width)
+        window, area = areas_3d[-1]
+
+        ## Return first found 3D area
+        # window, area = next(((w, a) for w in bpy.context.window_manager.windows for a in w.screen.areas if a.type == 'VIEW_3D'), (None,None))
+        # if not area:
+        #     print('Could not found area ! skip "set sidebar" operation')
+        #     return
+
+    sidebar = None
+
+    ## Set sidebar visibility
+    space_data = context.space_data
+    if not space_data:
+        space_data = area.spaces.active
+
+    if prefs.show_sidebar != 'NONE':
+        if space_data.show_region_ui and prefs.show_sidebar == 'HIDE':
+            space_data.show_region_ui = False
+        if not space_data.show_region_ui and prefs.show_sidebar == 'SHOW':
+            space_data.show_region_ui = True
+
+    ## Set sidebar panel tab
+    if bpy.app.version >= (4,2,0) and prefs.set_sidebar_tab and space_data.show_region_ui:
+        tab = prefs.sidebar_tab_target
+        if not tab.strip():
+            tab = 'Storytools'
+
+        ## 'active_panel_category' is readonly at first (Then goes to false after operator has finished)
+        ## Not working: Try to refresh UI to fix that once sidebar is opened
+        # context.area.regions.update()
+        # context.screen.update_tag()
+        if sidebar := next((r for r in area.regions if r.type == 'UI'), None):
+            with bpy.context.temp_override(window=window, area=area, region=sidebar):
+                try:
+                    # For now, just bypass the error, a second call works after full interface reload...
+                    sidebar.active_panel_category = tab
+                    sidebar.tag_redraw()
+                    print('sidebar category set to:', tab)
+                except AttributeError:
+                    print('Could not set sidebar category (retry can work)')
+                    pass
+
+        # print('finished', sidebar)
+""" Previous version
+def set_sidebar(area=None):
     if not area:
         area = bpy.context.area
     prefs = fn.get_addon_prefs()
     context = bpy.context
+
     sidebar = None
     # with bpy.context.temp_override(area=area):
     ## Set sidebar visibility
@@ -46,8 +107,13 @@ def set_sidebar(area=None):
                     print('Could not set sidebar category (retry can work)')
                     pass
         # print('finished', sidebar)
+ """
 
-def activate_workspace(name='Storyboard', context=None):
+def activate_workspace(name='Storyboard', filepath=None, context=None):
+    """Activate workspace by workspace name
+    filepath: if specified, fetch the workspace from this path
+    """
+
     if context is None:
         context = bpy.context
 
@@ -60,8 +126,10 @@ def activate_workspace(name='Storyboard', context=None):
         # return True
     else:
         # Same name with spaces as underscore
-        dir_name = name.replace(' ', '_')
-        filepath = APP_TEMPLATES_DIR / dir_name / 'startup.blend'
+        if filepath is None:    
+            ## use name
+            dir_name = name.replace(' ', '_')
+            filepath = APP_TEMPLATES_DIR / dir_name / 'startup.blend'
         
         ret = bpy.ops.workspace.append_activate(idname=name, filepath=str(filepath))
         if ret != {'FINISHED'}:
@@ -74,22 +142,119 @@ def activate_workspace(name='Storyboard', context=None):
     ## Get biggest viewport and set category toolbar to Storytools:
     return context.window.workspace
 
+## Load directly a single workspace
 class STORYTOOLS_OT_set_storyboard_workspace(bpy.types.Operator):
     bl_idname = "storytools.set_storyboard_workspace"
     bl_label = 'Set Storyboard Workspace'
-    bl_description = "Set story board workspace, import if needed"
+    bl_description = "Set storyboard workspace, import if needed"
     bl_options = {'REGISTER', 'INTERNAL'}
+
+    # workspace_name : bpy.props.BoolProperty(name='Storyboard', default=False)
 
     def execute(self, context):
         # ret = bpy.ops.workspace.append_activate(idname=name, filepath=str(filepath))
         activate_workspace(context=context)
-        areas_3d = [a for a in bpy.context.screen.areas if a.type == 'VIEW_3D']
-        if not areas_3d:
-            return {"FINISHED"}
-        areas_3d.sort(key=lambda x: x.width)
-        set_sidebar(areas_3d[-1])
+        ## Search in current screen
+        # problem : after activating new workpace, the context is lost (or half voided)
+        # areas_3d = [a for a in bpy.context.screen.areas if a.type == 'VIEW_3D']
+        # if not areas_3d:
+        #     return {"FINISHED"}
+        # areas_3d.sort(key=lambda x: x.width)
+        # set_sidebar(areas_3d[-1])
+        # set_sidebar()
+        bpy.app.timers.register(set_sidebar, first_interval=0.8)
         return {"FINISHED"}
 
+class STORYTOOLS_OT_set_storyboard_dual_window_workspace(bpy.types.Operator):
+    bl_idname = "storytools.set_storyboard_dual_window_workspace"
+    bl_label = 'Set Storyboard Dual Window Workspace'
+    bl_description = "Set Dual Window storyboard workspace to be able to work with storypencil"
+    bl_options = {'REGISTER', 'INTERNAL'}
+
+    def execute(self, context):
+        # name = "Board"
+        # filepath = str(DUAL_STORYBOARD_TEMPLATE_BLEND)
+        
+        # ret = bpy.ops.workspace.append_activate(idname=name, filepath=filepath)
+        # if ret != {'FINISHED'}:
+        #     print(f'Could not found "{name}" at {filepath}')
+        #     message = [f'Could not found "{name}" workspace at:',
+        #             str(filepath)]
+        #     fn.show_message_box(_message=message, _title='Workspace Not found', _icon='ERROR')
+        #     # return False
+
+        ## Setup current window workspace (Storyboard)
+        activate_workspace(context=context)
+        bpy.app.timers.register(set_sidebar, first_interval=0.8)
+
+        current_win = context.window
+        current_win_id = current_win.as_pointer()
+        
+        ## Get or create secondary window
+        if len(context.window_manager.windows) == 1:
+            bpy.ops.wm.window_new_main()
+            ## new window not active
+
+        ## Activate Sequencer on secondary window
+        if secondary_win := next((w for w in context.window_manager.windows if w.as_pointer() != current_win_id), None):
+            with context.temp_override(window=secondary_win):
+                ## load from storyboard template
+                activate_workspace('Sequencer', filepath=STORYBOARD_TEMPLATE_BLEND)
+                ## Can add some spark setup here (as option if an arg ops is passed)
+
+        return {"FINISHED"}
+
+## UNUSED CURRENTLY
+# -------------------------------------------------------------
+#  Setup Spark (spa sequencer) session (WIP, broken), better to integrate in aboce operator
+#
+# -------------------------------------------------------------
+
+"""
+class STORYTOOLS_OT_setup_spark(bpy.types.Operator):
+    bl_idname = "storytools.setup_spark"
+    bl_label = "Setup Dual Workspace For Spark"
+    bl_description = "Configure settings for a storyboard session using spark (spa-sequencer) for shot management"
+    bl_options = {'REGISTER', 'UNDO'}
+
+    @classmethod
+    def poll(cls, context):
+        return True
+
+    def get_workspace(self, type):
+        for wrk in bpy.data.workspaces:
+            if wrk.name == type:
+                return wrk
+
+        return None
+
+    def execute(self, context):
+        scene_base = context.scene
+
+        wk_storyboard = activate_workspace(name='Storyboard', context=context)
+        wk_sequencer = activate_workspace(name='Sequencer', context=context)
+        # Create Workspace
+        ## Add video editing
+        # template_path = None
+        # if "Video Editing" not in bpy.data.workspaces:
+        #     template_path = next(bpy.utils.app_template_paths(), None)
+        #     filepath = str(Path(template_path, "Video_Editing", "startup.blend"))
+        #     bpy.ops.workspace.append_activate(
+        #         idname="Video Editing", filepath=filepath)
+        
+        # Get/Create New scene
+
+        ## Using data
+        scene_edit = bpy.data.scenes.get('EDIT')
+        if not scene_edit:
+            scene_edit = bpy.data.scenes.get('Edit')
+        if not scene_edit:
+            scene_edit = bpy.data.scenes.new('EDIT')
+
+        # scene_edit.storypencil_edit_workspace = self.get_workspace("Storyboard")
+        bpy.app.timers.register(set_workpaces_on_dual_window_spark, first_interval=1.5)
+        return {"FINISHED"}
+"""
 
 ## UNUSED CURRENTLY
 # -------------------------------------------------------------
@@ -288,7 +453,7 @@ class STORYTOOLS_OT_setup_storypencil(bpy.types.Operator):
         
         # Append custom workspaces
         wk_storyboard = activate_workspace(name='Storyboard', context=context)
-        wk_sequencer = activate_workspace(name='Sequencer', context=context)
+        wk_sequencer = activate_workspace(name='Sequencer', filepath=STORYBOARD_TEMPLATE_BLEND, context=context)
 
         # Create Workspace
         template_path = None
@@ -357,7 +522,9 @@ class STORYTOOLS_OT_setup_storypencil(bpy.types.Operator):
 
 classes=(
     STORYTOOLS_OT_set_storyboard_workspace,
-    STORYTOOLS_OT_setup_storypencil,
+    STORYTOOLS_OT_set_storyboard_dual_window_workspace, # wip broken
+    # STORYTOOLS_OT_setup_spark, # wip broken
+    # STORYTOOLS_OT_setup_storypencil,
 )
 
 def register():